Reorganize datasets test modules
